requirements.txt
mysql_connector_repackaged==0.3.1
selenium==4.7.2
